feat: start to create ChannelsSlice
